Ore Sample Fixes (#92)
* Adjust ore sample's name

* Allow harvestTool to be set as null

* Extend off SubBlockBase instead of SubBlockPart for Ore Sample

I wanted to just set the ideal defaults for something like an ore sample, so I think this was the best path to go...

* Change Ore Sample material to GROUND

* Allow for floats to be used for hardness

* Revert "Extend off SubBlockBase instead of SubBlockPart for Ore Sample"

This reverts commit a62c285

* Set defaults for Ore Sample the better way.

* Cleanup - Mostly utilizing MaterialPartData.getValue

* Fix incorrect default variable being used for activatedText

+++ src/main/java/com/teamacronymcoders/base/subblocksystem/blocks/MissingSubBlock.java
@@ -42,7 +42,7 @@ public class MissingSubBlock implements ISubBlock {
     }
 
     @Override
-    public int getHardness() {
+    public float getHardness() {
         return 0;
     }
